Split out getDeserializedPostsQueryDetails
