+++ client/state/posts/utils/index.js
@@ -40,12 +40,6 @@ import stripHtml from 'lib/post-normalizer/rule-strip-html';
 import { getSerializedPostsQuery } from './get-serialized-posts-query';
 
 /**
- * Constants
- */
-
-const REGEXP_SERIALIZED_QUERY = /^((\d+):)?(.*)$/;
-
-/**
  * Utility
  */
 
@@ -62,27 +56,7 @@ const normalizeDisplayFlow = flow( [
 
 export { getNormalizedPostsQuery } from './get-normalized-posts-query';
 export { getSerializedPostsQuery } from './get-serialized-posts-query';
-
-/**
- * Returns an object with details related to the specified serialized query.
- * The object will include siteId and/or query object, if can be parsed.
- *
- * @param  {string} serializedQuery Serialized posts query
- * @returns {object}                 Deserialized posts query details
- */
-export function getDeserializedPostsQueryDetails( serializedQuery ) {
-	let siteId, query;
-
-	const matches = serializedQuery.match( REGEXP_SERIALIZED_QUERY );
-	if ( matches ) {
-		siteId = Number( matches[ 2 ] ) || undefined;
-		try {
-			query = JSON.parse( matches[ 3 ] );
-		} catch ( error ) {}
-	}
-
-	return { siteId, query };
-}
+export { getDeserializedPostsQueryDetails } from './get-deserialized-posts-query-details';
 
 /**
  * Returns a serialized posts query, excluding any page parameter
